refactor(benchmark): extract audit internals into fynd-tools-common
Move EthCallRunner, the aggregator quote model, BPS math, and the
FyndAggregator wrapper out of the audit subcommand and into
fynd-tools-common. Delete the duplicated erc20.rs and rely on
erc20-overrides via the shared EthCallRunner.

Split the monolithic audit.rs into focused modules (args, output,
execute, blocks, summary, mod) and break the oversized run,
execute_trade, and print_summary functions into smaller helpers.
aggregator.rs now holds only the external clients (Nordstern, KyberSwap,
0x). Behaviour is unchanged.
