+++ tools/benchmark/src/aggregator.rs
@@ -1,110 +1,16 @@
-//! External aggregator clients for Fynd audit comparisons.
+//! External aggregator clients for Fynd audit comparisons (Nordstern, KyberSwap, 0x).
 //!
 //! Add a new aggregator by implementing [`AggregatorClient`] and constructing it in
-//! `audit::run`.
+//! `audit::run`. The shared quote model (`AggregatorClient`, `AggregatorQuote`, …) lives in
+//! `fynd-tools-common`.
 
-use std::{collections::HashMap, fmt, time::Instant};
+use std::{collections::HashMap, time::Instant};
 
 use async_trait::async_trait;
+use fynd_tools_common::aggregator::{
+    AggregatorCalldata, AggregatorClient, AggregatorQuote, AggregatorStatus,
+};
 use serde::{Deserialize, Serialize};
-
-/// Quote status returned by an external aggregator.
-///
-/// Use this enum rather than comparing status strings directly.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AggregatorStatus {
-    Success,
-    NoAmount,
-    NoRoute,
-    /// The aggregator returned an HTTP error response (4xx/5xx).
-    HttpError {
-        code: u16,
-        snippet: String,
-    },
-}
-
-impl fmt::Display for AggregatorStatus {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Success => f.write_str("success"),
-            Self::NoAmount => f.write_str("no_amount"),
-            Self::NoRoute => f.write_str("no_route"),
-            Self::HttpError { code, snippet } => write!(f, "http_{code}: {snippet}"),
-        }
-    }
-}
-
-impl Serialize for AggregatorStatus {
-    fn serialize<S: serde::Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
-        s.serialize_str(&self.to_string())
-    }
-}
-
-/// On-chain transaction calldata returned by an aggregator for execution validation.
-#[derive(Debug, Clone, Serialize)]
-pub struct AggregatorCalldata {
-    /// Router / settlement contract address (hex with 0x prefix).
-    pub to: String,
-    /// ABI-encoded calldata (hex with 0x prefix).
-    pub data: String,
-    /// Native ETH value to send with the transaction (decimal string; usually `"0"`).
-    pub value: String,
-}
-
-/// A quote response from an external aggregator.
-#[derive(Debug, Clone, Serialize)]
-pub struct AggregatorQuote {
-    pub status: AggregatorStatus,
-    /// Output amount as a decimal string (token units).
-    pub amount_out: Option<String>,
-    /// Output amount net of gas costs (token units). Only populated by Fynd; `None` for all
-    /// other aggregators. Used as the baseline when computing gas-adjusted bps diffs.
-    pub amount_out_net_gas: Option<String>,
-    /// Total estimated gas units across all route legs.
-    pub gas_units: Option<u64>,
-    /// DEX names used in the route (e.g. `["uniswap_v3"]`).
-    pub protocols: Vec<String>,
-    /// Number of independent parallel sub-routes the aggregator split the order across.
-    /// `1` means a single path (no splitting). `None` when the aggregator does not expose
-    /// route structure (e.g. 0x `/price`).
-    pub num_splits: Option<usize>,
-    /// Wall-clock time from request dispatch to response body received.
-    pub response_time_ms: u64,
-    /// Encoded on-chain transaction. Present only when `fetch_calldata` was `true` and the
-    /// aggregator has an encoding endpoint. `None` otherwise.
-    pub calldata: Option<AggregatorCalldata>,
-    /// Pool-level route: `[protocol, component_id]` pairs (Fynd only; `None` for aggregators).
-    pub route: Option<Vec<[String; 2]>>,
-}
-
-impl AggregatorQuote {
-    pub fn is_success(&self) -> bool {
-        self.status == AggregatorStatus::Success
-    }
-}
-
-/// Interface for any external DEX aggregator.
-#[async_trait]
-pub trait AggregatorClient: Send + Sync {
-    /// Short label used in output (e.g. `"nordstern"`).
-    fn name(&self) -> &str;
-
-    /// Request a sell-side quote. Returns `Err` for network or parse failures; routing
-    /// outcomes (no route, HTTP error from the aggregator) are encoded in the returned
-    /// quote's `status` field.
-    ///
-    /// When `wallet` is `Some(address)`, the implementation fetches the encoded on-chain
-    /// transaction (populating [`AggregatorQuote::calldata`]) and uses `address` as the
-    /// sender/recipient so that balance-delta validation reads the correct account.
-    /// Pass `None` when calldata is not needed.
-    async fn quote(
-        &self,
-        token_in: &str,
-        token_out: &str,
-        amount: &str,
-        wallet: Option<&str>,
-    ) -> anyhow::Result<AggregatorQuote>;
-}
 
 // ─── Nordstern Finance ────────────────────────────────────────────────────────
 
@@ -728,162 +634,9 @@ impl AggregatorClient for ZeroExClient {
     }
 }
 
-// ─── bps comparison helpers ───────────────────────────────────────────────────
-
-/// Compare baseline vs another participant on raw output amounts, in basis points.
-///
-/// Positive = baseline delivers more output. `None` when amounts are zero or unparseable.
-pub fn raw_bps_diff(baseline: &str, other: &str) -> Option<f64> {
-    let b: f64 = baseline
-        .parse()
-        .ok()
-        .filter(|&v: &f64| v > 0.0)?;
-    let o: f64 = other
-        .parse()
-        .ok()
-        .filter(|&v: &f64| v > 0.0)?;
-    Some((b - o) / o * 10_000.0)
-}
-
-/// Compare baseline vs another participant net-of-gas, in basis points.
-///
-/// Derives the token-per-gas-unit price from the baseline's `(raw - net_gas) / gas_units`,
-/// then applies it to the other participant's gas units. Returns `None` when either gas
-/// figure is zero, when `baseline_net_gas` is absent, or when gas exceeds 5% of output.
-/// Call this twice with reported and on-chain gas values to keep the two measurements
-/// independent.
-pub fn gas_adjusted_bps_diff(
-    baseline_raw: &str,
-    baseline_net_gas: Option<&str>,
-    baseline_gas_units: u64,
-    other_raw: &str,
-    other_gas_units: u64,
-) -> Option<f64> {
-    let baseline_net_gas = baseline_net_gas?;
-    if baseline_gas_units == 0 || other_gas_units == 0 {
-        return None;
-    }
-    let b_raw: f64 = baseline_raw
-        .parse()
-        .ok()
-        .filter(|&v: &f64| v > 0.0)?;
-    let b_net: f64 = baseline_net_gas
-        .parse()
-        .ok()
-        .filter(|&v: &f64| v > 0.0)?;
-    let o_raw: f64 = other_raw
-        .parse()
-        .ok()
-        .filter(|&v: &f64| v > 0.0)?;
-
-    let gas_cost = b_raw - b_net;
-
-    // Skip trades where gas exceeds 5% of output — uneconomical and the bps math becomes noise.
-    if gas_cost * 20.0 > b_raw {
-        return None;
-    }
-
-    let token_per_gas = gas_cost / baseline_gas_units as f64;
-    let o_net = o_raw - (other_gas_units as f64 * token_per_gas);
-
-    if o_net <= 0.0 {
-        return None;
-    }
-    Some((b_net - o_net) / o_net * 10_000.0)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn raw_bps_diff_fynd_better() {
-        let d = raw_bps_diff("10100", "10000").unwrap();
-        assert!((d - 100.0).abs() < 0.01, "expected 100 bps, got {d}");
-    }
-
-    #[test]
-    fn raw_bps_diff_agg_better() {
-        let d = raw_bps_diff("9900", "10000").unwrap();
-        assert!((d - (-100.0)).abs() < 0.01, "expected -100 bps, got {d}");
-    }
-
-    #[test]
-    fn raw_bps_diff_equal() {
-        assert_eq!(raw_bps_diff("10000", "10000"), Some(0.0));
-    }
-
-    #[test]
-    fn raw_bps_diff_zero_agg() {
-        assert_eq!(raw_bps_diff("10000", "0"), None);
-    }
-
-    #[test]
-    fn raw_bps_diff_invalid() {
-        assert_eq!(raw_bps_diff("abc", "10000"), None);
-    }
-
-    #[test]
-    fn aggregator_status_is_success_enum_match() {
-        let q = AggregatorQuote {
-            status: AggregatorStatus::Success,
-            amount_out: Some("1000".to_string()),
-            amount_out_net_gas: None,
-            gas_units: None,
-            protocols: vec![],
-            num_splits: None,
-            response_time_ms: 100,
-            calldata: None,
-            route: None,
-        };
-        assert!(q.is_success());
-    }
-
-    #[test]
-    fn aggregator_status_no_amount_not_success() {
-        let q = AggregatorQuote {
-            status: AggregatorStatus::NoAmount,
-            amount_out: None,
-            amount_out_net_gas: None,
-            gas_units: None,
-            protocols: vec![],
-            num_splits: None,
-            response_time_ms: 100,
-            calldata: None,
-            route: None,
-        };
-        assert!(!q.is_success());
-    }
-
-    #[test]
-    fn aggregator_status_http_error_not_success() {
-        let q = AggregatorQuote {
-            status: AggregatorStatus::HttpError { code: 429, snippet: "rate limited".to_string() },
-            amount_out: None,
-            amount_out_net_gas: None,
-            gas_units: None,
-            protocols: vec![],
-            num_splits: None,
-            response_time_ms: 50,
-            calldata: None,
-            route: None,
-        };
-        assert!(!q.is_success());
-    }
-
-    #[test]
-    fn aggregator_status_serialises_to_string() {
-        assert_eq!(serde_json::to_string(&AggregatorStatus::Success).unwrap(), r#""success""#);
-        assert_eq!(serde_json::to_string(&AggregatorStatus::NoRoute).unwrap(), r#""no_route""#);
-        assert_eq!(
-            serde_json::to_string(&AggregatorStatus::HttpError {
-                code: 429,
-                snippet: "rate limited".to_string()
-            })
-            .unwrap(),
-            r#""http_429: rate limited""#
-        );
-    }
 
     #[test]
     fn kyberswap_route_summary_round_trips() {
